fight_classifier/requirements.txt
absl-py==1.2.0
aiobotocore==2.4.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.2.0
albumentations==1.3.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
daemonize==2.5.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distlib==0.3.6
distro==1.7.0
dpath==2.0.6
dulwich==0.20.46
dvc==2.28.0
dvc-data==0.13.0
dvc-http==2.27.2
dvc-objects==0.5.0
dvc-render==0.0.11
dvc-s3==2.20.1
dvc-task==0.1.2
dvclive==0.11.0
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
-e git+ssh://git@github.com/ernest-tg/hiring-challenge.git@08e2e6690d1c8e30baf0c8ba3f3ffcdfe12ddb21#egg=fight_classifier&subdirectory=fight_classifier
filelock==3.8.0
fire==0.4.0
flake8==5.0.4
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grandalf==0.6
grpcio==1.49.1
guildai==0.8.1
hydra-core==1.2.0
idna==3.4
imageio==2.22.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iterative-telemetry==0.0.5
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
multidict==6.0.2
mypy==0.981
mypy-extensions==0.4.3
nanotime==0.5.2
natsort==8.2.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
networkx==2.8.7
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.1
omegaconf==2.2.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
pydot==1.4.2
pyflakes==2.5.0
pygit2==1.10.1
Pygments==2.13.0
pygtrie==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3fs==2022.8.2
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-optimize==0.9.0
scikit-video==1.1.11
scipy==1.9.1
scmrepo==0.1.1
seaborn==0.12.0
Send2Trash==1.8.0
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.1
tabulate==0.8.10
tabview==1.4.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
zc.lockfile==2.0
